заполнение шаблона данными Дз к 7 уроку часть 1
