+++ js/main.js
@@ -1,1 +1,2 @@
 import './data.js';
+import './make-fotos.js';
