--- js/utils/make-id-generator.js
@@ -1,6 +1,0 @@
-const makeIdGenerator = () => {
-  let id = 0;
-  return () => { id = id + 1; return id; };
-};
-
-export { makeIdGenerator };
